refactor: rename and reorganize backend and frontend packages, update application names, and adjust HTML fragment references
