--- backend/.vscode/launch.json
@@ -1,28 +1,0 @@
-{
-  "configurations": [
-    {
-      "type": "java",
-      "name": "NotesApplication",
-      "request": "launch",
-      "mainClass": "NotesApplication",
-      "projectName": "sample"
-    },
-    {
-      "type": "java",
-      "name": "NotesApplication",
-      "request": "launch",
-      "mainClass": "io.d4rkr0n1n.sample.NotesApplication",
-      "projectName": "sample"
-    },
-    {
-      "type": "java",
-      "name": "Spring Boot-SampleApplication<sample>",
-      "request": "launch",
-      "cwd": "${workspaceFolder}",
-      "mainClass": "io.d4rkr0n1n.sample.SampleApplication",
-      "projectName": "sample",
-      "args": "",
-      "envFile": "${workspaceFolder}/.env"
-    }
-  ]
-}
