+++ backend/src/main/java/io/d4rkr0n1n/backend/Application.java
@@ -1,13 +1,13 @@
-package io.d4rkr0n1n.sample;
+package io.d4rkr0n1n.backend;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 @SpringBootApplication
-public class NotesBackendApplication {
+public class Application {
 
 	public static void main(String[] args) {
-		SpringApplication.run(NotesBackendApplication.class, args);
+		SpringApplication.run(Application.class, args);
 	}
 
 }
