+++ backend/src/main/java/io/d4rkr0n1n/backend/service/NotesResponseService.java
@@ -1,4 +1,4 @@
-package io.d4rkr0n1n.sample.service;
+package io.d4rkr0n1n.backend.service;
 
 import java.util.List;
 import java.util.UUID;
@@ -6,8 +6,8 @@ import java.util.UUID;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
-import io.d4rkr0n1n.sample.helper.ResponseHelper;
-import io.d4rkr0n1n.sample.model.Note;
+import io.d4rkr0n1n.backend.helper.ResponseHelper;
+import io.d4rkr0n1n.backend.model.Note;
 
 @Service
 public class NotesResponseService {
