+++ frontend/src/main/java/io/d4rkr0n1n/frontend/Application.java
@@ -1,4 +1,4 @@
-package io.d4rkr0n1n.sample;
+package io.d4rkr0n1n.frontend;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
@@ -6,10 +6,10 @@ import org.springframework.cloud.openfeign.EnableFeignClients;
 
 @SpringBootApplication
 @EnableFeignClients
-public class NotesFrontendApplication {
+public class Application {
 
 	public static void main(String[] args) {
-		SpringApplication.run(NotesFrontendApplication.class, args);
+		SpringApplication.run(Application.class, args);
 	}
 
 }
