+++ frontend/src/main/java/io/d4rkr0n1n/frontend/model/Note.java
@@ -1,22 +1,17 @@
-package io.d4rkr0n1n.sample.model;
+package io.d4rkr0n1n.frontend.model;
 
 import java.sql.Timestamp;
 import java.util.UUID;
 
-import jakarta.persistence.Entity;
-import jakarta.persistence.Id;
-import jakarta.persistence.Table;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-@Table
-@Entity
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
 public class Note {
-  @Id UUID id;
+  UUID id;
   String name;
   String contents;
   Timestamp timestamp;
