+++ frontend/src/main/resources/templates/index.html
@@ -22,7 +22,7 @@
     <hr>
 
   </div>
-  <div th:replace="fragments/notesList :: notes"></div>
+  <div th:replace="~{fragments/notesList :: notes}"></div>
   <script th:src="@{/js/script.js}"></script>
   <script src="/webjars/bootstrap/5.3.3/js/bootstrap.bundle.min.js"></script>
 </body>
